+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
 v0.88.2
 
+- Expose `useH2ForAuthChat` remote configuration key to use HTTP/2 for AuthenticatedChatConnection's non-fronted connections.
+
 - The `disableNagleAlgorithm` remote config flag has been removed, as the experiment has been deployed successfully.